Move per-DatasetType collection logic down to StandardRepoConverter.
This lets us move some datasets to their own RUN collection during
conversion rather than forcing us to put them in the main RUN for that
Gen2 rerun.  That's useful in at least ci_hsc, where we have to put
the brightObjectMasks in the rerun in Gen2 because that's where the
skymap is defined, but we'd rather put them in a top-level collection
in Gen3.